feat: filters logic added
