+++ composables/filters/useFilterOptions.ts
@@ -1,0 +1,16 @@
+import type { SelectOption } from '~/types/filters'
+
+export const useFilterOptions = () => {
+  const defaultSortOptions: SelectOption[] = [
+    { value: 'price-asc', label: 'Price: Low to High' },
+    { value: 'price-desc', label: 'Price: High to Low' },
+    { value: 'name-asc', label: 'Name: A to Z' },
+    { value: 'name-desc', label: 'Name: Z to A' },
+    { value: 'newest', label: 'Newest First' },
+    { value: 'rating-desc', label: 'Highest Rated' },
+  ]
+
+  return {
+    defaultSortOptions,
+  }
+}
