+++ pages/catalogue.vue
@@ -2,58 +2,29 @@
   import { useGetAllProducts } from '@/composables/api/products/useGetAllProducts'
   import { useNotification } from '~/composables/notification/useNotification'
   import { usePagination } from '~/composables/usePagination'
+  import { useProductFilters } from '~/composables/filters/useProductFilters'
+  import { useProductFiltering } from '~/composables/filters/useProductFiltering'
+  import { useProductCategories } from '~/composables/filters/useProductCategories'
+  import type { FiltersState } from '~/types/filters'
   import { debounce } from '~/utils/debounce'
-  import { useRoute, useRouter } from 'nuxt/app'
 
   const { showError } = useNotification()
-  const route = useRoute()
-  const router = useRouter()
 
-  export interface FiltersState {
-    search: string
-    category: string
-    sortBy: string
-    minPrice: number
-    maxPrice: number
-    onSale: boolean
-    inStock: boolean
-  }
-
-  // Состояние фильтров с явным типом
-  const filters = reactive<FiltersState>({
-    search: '',
-    category: '',
-    sortBy: '',
-    minPrice: 0,
-    maxPrice: 500,
-    onSale: false,
-    inStock: false,
-  })
-
-  // Загрузка всех продуктов через ваш useGetAllProducts
   const {
     data: allProducts,
     pending: allProductsPending,
     error: allProductsError,
   } = useGetAllProducts()
 
-  // Загрузка категорий из продуктов
-  const categories = computed(() => {
-    if (!allProducts.value) return []
+  const { filters, initFilters, updateQueryParams } = useProductFilters()
+  const { categories } = useProductCategories(allProducts)
+  const { getFilteredProducts } = useProductFiltering(allProducts)
+  const filteredProducts = computed(() => getFilteredProducts(filters))
 
-    const uniqueCategories = [...new Set(allProducts.value.map((product) => product.category))]
-    return uniqueCategories.map((category) => ({
-      value: category,
-      label: category.charAt(0).toUpperCase() + category.slice(1).replace(/'/g, "'"),
-    }))
-  })
-
-  // Чтение query параметров при загрузке
   onMounted(() => {
-    updateFiltersFromQuery()
+    initFilters()
   })
 
-  // Синхронизация фильтров с URL - явно указываем тип
   watch(
     filters,
     debounce((newFilters: FiltersState) => {
@@ -62,7 +33,6 @@
     { deep: true },
   )
 
-  // Обработка ошибок
   watch(allProductsError, (newError) => {
     if (newError) {
       console.error('Error loading products:', newError)
@@ -70,99 +40,6 @@
     }
   })
 
-  // Обновление фильтров из query параметров
-  const updateFiltersFromQuery = () => {
-    const query = route.query
-
-    if (query.search) filters.search = String(query.search)
-    if (query.category) filters.category = String(query.category)
-    if (query.sortBy) filters.sortBy = String(query.sortBy)
-    if (query.minPrice) filters.minPrice = Number(query.minPrice)
-    if (query.maxPrice) filters.maxPrice = Number(query.maxPrice)
-    if (query.onSale) filters.onSale = query.onSale === 'true'
-    if (query.inStock) filters.inStock = query.inStock === 'true'
-  }
-
-  // Обновление query параметров
-  const updateQueryParams = (newFilters: FiltersState) => {
-    const query: Record<string, string> = {}
-
-    if (newFilters.search) query.search = newFilters.search
-    if (newFilters.category) query.category = newFilters.category
-    if (newFilters.sortBy) query.sortBy = newFilters.sortBy
-    if (newFilters.minPrice > 0) query.minPrice = newFilters.minPrice.toString()
-    if (newFilters.maxPrice < 500) query.maxPrice = newFilters.maxPrice.toString()
-    if (newFilters.onSale) query.onSale = newFilters.onSale.toString()
-    if (newFilters.inStock) query.inStock = newFilters.inStock.toString()
-
-    router.replace({ query })
-  }
-
-  // Фильтрация и сортировка продуктов
-  const filteredProducts = computed(() => {
-    if (!allProducts.value) return []
-
-    let filtered = [...allProducts.value]
-
-    // Фильтр по категории
-    if (filters.category) {
-      filtered = filtered.filter((product) => product.category === filters.category)
-    }
-
-    // Фильтр по поиску
-    if (filters.search) {
-      const searchLower = filters.search.toLowerCase()
-      filtered = filtered.filter(
-        (product) =>
-          product.title.toLowerCase().includes(searchLower) ||
-          product.description.toLowerCase().includes(searchLower),
-      )
-    }
-
-    // Фильтр по цене
-    filtered = filtered.filter(
-      (product) => product.price >= filters.minPrice && product.price <= filters.maxPrice,
-    )
-
-    // Фильтр по наличию
-    if (filters.inStock) {
-      filtered = filtered.filter((product) => product.rating?.count && product.rating.count > 0)
-    }
-
-    // Фильтр по распродаже
-    if (filters.onSale) {
-      // Товары со скидкой (цена меньше 50)
-      filtered = filtered.filter((product) => product.price < 50)
-    }
-
-    // Сортировка
-    if (filters.sortBy) {
-      switch (filters.sortBy) {
-        case 'price-asc':
-          filtered.sort((a, b) => a.price - b.price)
-          break
-        case 'price-desc':
-          filtered.sort((a, b) => b.price - a.price)
-          break
-        case 'name-asc':
-          filtered.sort((a, b) => a.title.localeCompare(b.title))
-          break
-        case 'name-desc':
-          filtered.sort((a, b) => b.title.localeCompare(a.title))
-          break
-        case 'newest':
-          filtered.sort((a, b) => b.id - a.id)
-          break
-        case 'rating-desc':
-          filtered.sort((a, b) => (b.rating?.rate || 0) - (a.rating?.rate || 0))
-          break
-      }
-    }
-
-    return filtered
-  })
-
-  // Пагинация
   const {
     currentPage,
     totalPages,
@@ -172,7 +49,6 @@
     itemsPerPage: 6,
   })
 
-  // Мобильные фильтры
   const isShowMobFilters = ref(false)
 
   const toggleShowMobFilters = () => {
@@ -183,12 +59,10 @@
     isShowMobFilters.value = false
   }
 
-  // Обработчик изменения фильтров
   const handleFiltersChange = (newFilters: FiltersState) => {
     Object.assign(filters, newFilters)
   }
 
-  // Сброс текущей страницы при изменении фильтров
   watch(
     filters,
     () => {
